Seperated definitions from implementation.
