+++ Client-Project/CMakeLists.txt
@@ -6,4 +6,4 @@ cmake_minimum_required (VERSION 3.8)
 project ("Client-Project")
 
 # Add source to this project's executable.
-add_executable (Client "client.cpp")
+add_executable (Client "main.cpp")
