+++ Client-Project/client.cpp
@@ -83,72 +83,3 @@ bool ConnectToServer(const char* serverAddress, int port, int& sock)
 
     return true;
 }
-
-/**
- * @brief This is the entry for the client program.
- * 
- * @param argc 
- * @param argv 
- * @return int 
- */
-int main(int argc, char* argv[])
-{
-    int sock = 0;
-    
-    const char* connectRPC = "connect;MIKE;MIKE;";
-    //const char* statusRPC = "status;";
-    const char* logoffRPC = "disconnect;";
-    char buffer[1024] = { 0 };
-    const char* serverAddress = argv[1];
-    const int port = atoi(argv[2]);
-
-    bool bConnect = ConnectToServer(serverAddress, port, sock);
-
-    if (bConnect)
-    {
-        strcpy(buffer, connectRPC);
-        int nlen = strlen(buffer);
-        buffer[nlen] = 0;   // Put the null terminator
-
-        send(sock, buffer, strlen(buffer) + 1, 0);
-        read(sock, buffer, 1024);
-
-        read(sock, buffer, 1024);
-        printf("%s\n", buffer);
-
-        printf("What is your message for the server? ");
-        char message[80];
-        std::cin.getline(message, sizeof message);
-        send(sock, message, strlen(message) + 1, 0);
-    }
-    else
-    {
-        printf("\nExit without calling RPC.\n");
-    }
-
-    // Sleep 1 to 10 seconds randomly
-    int seconds = 1 + (rand() % static_cast<int>(10));
-    printf("\nSleeping...\n");
-    sleep(seconds);
-
-    // Disconnect Message
-    if (bConnect)
-    {
-        strcpy(buffer, logoffRPC);
-        int nlen = strlen(buffer);
-        buffer[nlen] = 0;   // Put the null terminator
-        send(sock, buffer, strlen(buffer) + 1, 0);
-        read(sock, buffer, 1024);
-
-        printf("\nDisconnected.\n");
-    }
-    else
-    {
-        printf("\nExit without calling RPC.\n");
-    }
-
-    // Terminate connection.
-    close(sock);
-
-    return 0;
-}
